created common resources folder in client. moved methods.js there
