+++ src/client/src/resources/methods.js
